chore: migrate collab storage to appflowy-collaborate
