+++ services/appflowy-collaborate/Cargo.toml
@@ -54,6 +54,8 @@ redis = "0.25.2"
 parking_lot = "0.12.1"
 lazy_static = "1.4.0"
 itertools = "0.12.0"
+validator = "0.16.1"
+workspace-access.workspace = true
 
 [dev-dependencies]
 rand = "0.8.5"
