+++ services/appflowy-collaborate/src/collab/mod.rs
@@ -3,3 +3,8 @@ pub mod cache;
 pub mod disk_cache;
 pub mod mem_cache;
 pub mod notification;
+pub mod queue;
+mod queue_redis_ops;
+pub mod storage;
+pub mod validator;
+pub use queue_redis_ops::{PendingWrite, RedisSortedSet, WritePriority};
