+++ services/appflowy-collaborate/src/snapshot/snapshot_control.rs
@@ -1,6 +1,6 @@
-use crate::biz::collab::metrics::CollabMetrics;
-use crate::biz::snapshot::cache::SnapshotCache;
-use crate::biz::snapshot::queue::PendingQueue;
+use crate::metrics::CollabMetrics;
+use crate::snapshot::cache::SnapshotCache;
+use crate::snapshot::queue::PendingQueue;
 use crate::state::RedisConnectionManager;
 use anyhow::anyhow;
 use app_error::AppError;
