+++ src/api/ws.rs
@@ -15,11 +15,11 @@ use app_error::AppError;
 use appflowy_collaborate::actix_ws::client::rt_client::RealtimeClient;
 use appflowy_collaborate::actix_ws::server::RealtimeServerActor;
 use appflowy_collaborate::collab::access_control::RealtimeCollabAccessControlImpl;
+use appflowy_collaborate::collab::storage::CollabAccessControlStorage;
 use collab_rt_entity::user::{AFUserChange, RealtimeUser, UserMessage};
 use collab_rt_entity::RealtimeMessage;
 use shared_entity::response::AppResponseError;
 
-use crate::biz::collab::storage::CollabAccessControlStorage;
 use crate::biz::user::auth::jwt::{authorization_from_token, UserUuid};
 use crate::state::AppState;
