+++ src/application.rs
@@ -9,12 +9,8 @@ use access_control::access::{enable_access_control, AccessControl};
 
 use crate::api::ai_tool::ai_tool_scope;
 use crate::api::chat::chat_scope;
-use crate::biz::collab::access_control::{
-  CollabMiddlewareAccessControl, CollabStorageAccessControlImpl,
-};
-use crate::biz::collab::storage::CollabStorageImpl;
+use crate::biz::collab::access_control::CollabMiddlewareAccessControl;
 use crate::biz::pg_listener::PgListeners;
-use crate::biz::snapshot::SnapshotControl;
 use crate::biz::workspace::access_control::WorkspaceMiddlewareAccessControl;
 use crate::config::config::{Config, DatabaseSetting, GoTrueSetting, S3Setting};
 use crate::middleware::access_control_mw::MiddlewareAccessControlTransform;
@@ -32,11 +28,13 @@ use anyhow::{Context, Error};
 use appflowy_ai_client::client::AppFlowyAIClient;
 use appflowy_collaborate::actix_ws::server::RealtimeServerActor;
 use appflowy_collaborate::collab::access_control::{
-  CollabAccessControlImpl, RealtimeCollabAccessControlImpl,
+  CollabAccessControlImpl, CollabStorageAccessControlImpl, RealtimeCollabAccessControlImpl,
 };
 use appflowy_collaborate::collab::cache::CollabCache;
+use appflowy_collaborate::collab::storage::CollabStorageImpl;
 use appflowy_collaborate::command::{CLCommandReceiver, CLCommandSender};
 use appflowy_collaborate::shared_state::RealtimeSharedState;
+use appflowy_collaborate::snapshot::SnapshotControl;
 use appflowy_collaborate::CollaborationServer;
 use database::file::bucket_s3_impl::S3BucketStorage;
 use openssl::ssl::{SslAcceptor, SslAcceptorBuilder, SslFiletype, SslMethod};
