+++ src/biz/mod.rs
@@ -1,7 +1,6 @@
 pub mod chat;
 pub mod collab;
 pub mod pg_listener;
-pub mod snapshot;
 pub mod user;
 pub mod utils;
 pub mod workspace;
